\name{tidyCpp-package}
\alias{tidyCpp-package}
\alias{tidyCpp}
\docType{package}
\title{\packageTitle{tidyCpp}}
\description{\packageDescription{tidyCpp}}
\section{Package Content}{\packageIndices{tidyCpp}}
\author{\packageAuthor{tidyCpp}}
\section{Maintainer}{\packageMaintainer{tidyCpp}}
\keyword{package}
